third_party/blink/web_tests/virtual/legacy-client-hints-no-fp-delegation/external/wpt/client-hints/accept-ch-stickiness/cross-origin-navigation.https-expected.txt

This is a testharness.js-based test.
PASS cross origin navigation precondition: Test that the browser does not have client hints preferences cached
PASS cross origin navigation set Accept-CH
FAIL cross origin navigation got client hints according to expectations. assert_equals: message from opened page expected "PASS" but got "FAIL"
Harness: the test ran to completion.
